tests/wpt/tests/css/css-layout-api/supports.https.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.css-houdini.org/css-layout-api/#valdef-display-layout">
<meta name="assert" content="This test checks that a layout() function works correctly with CSS.supports()." />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
test(function() {
  assert_true(CSS.supports('display', 'layout(foo)'));
});
</script>
